requirements.txt
# Telegram Bot Framework
aiogram==3.5.0

# Web Framework для Mini App
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
pytz==2023.3

# Production Server
gunicorn==21.2.0
